requirement.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
google-generativeai
faiss-cpu
selenium
webdriver-manager
